requirements.txt
ttkbootstrap
opencv-python
ultralytics
numpy
pillow
